+++ SMFL_01/SMFL_01/Entity.cpp
@@ -21,7 +21,7 @@ void Entity::CreateEntity(sf::Vector2f _pos, std::string _spritePath)
 
 	mVelocity = sf::Vector2f(0.0f, 0.0f);
 
-	mSize = sf::Vector2f(mTexture.getSize().x, mTexture.getSize().y);
+	mSize = sf::Vector2f(static_cast<float>(mTexture.getSize().x), static_cast<float>(mTexture.getSize().y));
 
 	mRect = sf::FloatRect(mPosition, mSize);
